tests/wpt/web-platform-tests/web-animations/interfaces/KeyframeEffect/copy-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>KeyframeEffect copy constructor tests</title>
<link rel="help"
href="https://w3c.github.io/web-animations/#dom-keyframeeffect-keyframeeffect-source">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
"use strict";

test(function(t) {
  var effect = new KeyframeEffectReadOnly(createDiv(t), null);
  assert_equals(effect.constructor.name, 'KeyframeEffectReadOnly');
  assert_equals(effect.timing.constructor.name,
                'AnimationEffectTimingReadOnly');

  // Make a mutable copy
  var copiedEffect = new KeyframeEffect(effect);
  assert_equals(copiedEffect.constructor.name, 'KeyframeEffect');
  assert_equals(copiedEffect.timing.constructor.name, 'AnimationEffectTiming');
}, 'Test mutable copy from a KeyframeEffectReadOnly source');

</script>
</body>
